Add partial organizations index on unique name when not deleted.
Organizations were previously indexed on their unique name; this caused an index constraint error when "Acme Museum" was deleted and then recreated. This partial index resolves this such that multiple deleted organizations with the same name can exist, but only a single "Acme Museum" that is not deleted can only ever exist.